+++ sdk/profile/src/main/java/io/opentelemetry/sdk/profile/data/AttributeUnitData.java
@@ -22,5 +22,5 @@ public interface AttributeUnitData {
   /**
    * Index into string table.
    */
-  int getUnitIndex();
+  long getUnitIndex();
 }
