+++ sdk/profile/src/main/java/io/opentelemetry/sdk/profile/data/FunctionData.java
@@ -25,20 +25,20 @@ public interface FunctionData {
    * Name of the function, in human-readable form if available.
    * Index into string table.
    */
-  int getNameIndex();
+  long getNameIndex();
 
   /**
    * Name of the function, as identified by the system.
    * For instance, it can be a C++ mangled name.
    * Index into string table.
    */
-  int getSystemNameIndex();
+  long getSystemNameIndex();
 
   /**
    * Source file containing the function.
    * Index into string table.
    */
-  int getFilenameIndex();
+  long getFilenameIndex();
 
   /**
    * Line number in source file.
